[www/starters] rearrange components to make it more readable
